refactor: share splitModelRef helper
